微服务实现商城后台系统---添加管理员登录功能

+++ springcloud-common/src/main/java/com/ljx/springcloud/utils/PageResult.java
@@ -6,7 +6,7 @@ public class PageResult<T> {
     //总条数
     private Long total;
     //总页数
-    private Long totalPage;
+    private Integer totalPage;
     //当前页数据
     private List<T> items;
 
@@ -18,7 +18,7 @@ public class PageResult<T> {
         this.items = items;
     }
 
-    public PageResult(Long total, Long totalPage, List<T> items) {
+    public PageResult(Long total, Integer totalPage, List<T> items) {
         this.total = total;
         this.totalPage = totalPage;
         this.items = items;
@@ -32,11 +32,11 @@ public class PageResult<T> {
         this.total = total;
     }
 
-    public Long getTotalPage() {
+    public Integer getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(Long totalPage) {
+    public void setTotalPage(Integer totalPage) {
         this.totalPage = totalPage;
     }
